Renamed object.hpp to window.hpp
