+++ src/gui/windows/registers_window.hpp
@@ -2,7 +2,7 @@
 
 #include <imgui/imgui.h>
 
-#include "../object.hpp"
+#include "../window.hpp"
 
 #include "../../core/arm1176jzf_s/cpu_core.hpp"
